Added controller functionality

+++ try_clap/src/movie.rs
@@ -2,12 +2,12 @@ use std::fmt::Display;
 
 pub struct Movie {
     pub name: String,
-    pub year: u8,
+    pub year: u16,
     pub rating: f32,
 }
 
 impl Movie {
-    pub fn new(name: String, year: u8, rating: f32) -> Self {
+    pub fn new(name: String, year: u16, rating: f32) -> Self {
         Movie { name, year, rating }
     }
 }
